+++ tests/BigCommerce/Customers/CustomerFormFieldValuesApiTest.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace BigCommerce\Tests\Customers;
+
+use BigCommerce\Tests\BigCommerceApiTest;
+
+class CustomerFormFieldValuesApiTest extends BigCommerceApiTest
+{
+    public function testCanGetCustomerFormFieldValues()
+    {
+        $this->markTestIncomplete();
+    }
+
+    public function testCanUpsertCustomerFormFieldValues()
+    {
+        $this->markTestIncomplete();
+    }
+}