+++ RELEASE_NOTES.md
@@ -5,4 +5,5 @@ Including:
 - Customer Addresses
 - Customer Attributes
 - Customer Attribute Values
+- Customer Form Field Values